requirements.txt
absl-py==2.0.0
astunparse==1.6.3
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deepface==0.0.79
dnspython==2.5.0
eventlet==0.35.1
filelock==3.13.1
fire==0.5.0
Flask==3.0.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
gast==0.5.4
gdown==4.7.1
gevent==24.2.1
gevent-websocket==0.10.1
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.59.3
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.15.0
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mtcnn==0.1.1
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.23.4
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.8.2
python-socketio==5.11.1
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
socketIO-client==0.7.2
soupsieve==2.5
SQLAlchemy==2.0.23
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.2
